Move LSP-related tests to match the new LSP plugin structure
