+++ python/tests/plugins/lsp_server/test_code_completion_provider.py
@@ -7,8 +7,8 @@ from aac.lang.definitions.search import search_definition
 from aac.plugins.lsp_server.providers.code_completion_provider import SPACE_TRIGGER
 
 from tests.helpers.lsp.responses.completion_response import CompletionResponse
-from tests.lang.lsp.base_lsp_test_case import BaseLspTestCase
-from tests.lang.lsp.definition_constants import TEST_DOCUMENT_NAME, TEST_PARTIAL_CONTENT
+from tests.plugins.lsp_server.base_lsp_test_case import BaseLspTestCase
+from tests.plugins.lsp_server.definition_constants import TEST_DOCUMENT_NAME, TEST_PARTIAL_CONTENT
 
 
 class TestCodeCompletionProvider(BaseLspTestCase, IsolatedAsyncioTestCase):
